Zero Point Module steam ready
