+++ Source/ModLoader.sln
@@ -40,7 +40,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuildingModifier", "Buildin
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NoDamage", "NoDamageMod\NoDamage.csproj", "{480FFA2E-7673-4A3E-994A-DC401B812D95}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZeroPointModuleMod", "ZeroPointModuleMod\ZeroPointModuleMod.csproj", "{1F1220E9-0CBD-4776-A9FA-31941CE63FA0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZeroPointModule", "ZeroPointModule\ZeroPointModule.csproj", "{1F1220E9-0CBD-4776-A9FA-31941CE63FA0}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WorldGenReloaded", "WorldGenReloadedMod\WorldGenReloaded.csproj", "{F648211E-4AFE-44B3-8A5E-B89D57CF34A7}"
 EndProject
@@ -309,7 +309,7 @@ Global
 		{4BF8EC3F-EB0C-4FC7-B2F5-458E91F63E96} = {F23F1C19-7F80-41CC-86F7-F720C6D3C43B}
 		{AE64BD07-9B27-4E16-9767-DD7894F70806} = {F23F1C19-7F80-41CC-86F7-F720C6D3C43B}
 		{480FFA2E-7673-4A3E-994A-DC401B812D95} = {F23F1C19-7F80-41CC-86F7-F720C6D3C43B}
-		{1F1220E9-0CBD-4776-A9FA-31941CE63FA0} = {F23F1C19-7F80-41CC-86F7-F720C6D3C43B}
+		{1F1220E9-0CBD-4776-A9FA-31941CE63FA0} = {8E4312B0-3AE8-42CB-A2ED-CB8E39B39CD4}
 		{F648211E-4AFE-44B3-8A5E-B89D57CF34A7} = {F23F1C19-7F80-41CC-86F7-F720C6D3C43B}
 		{0A8D7289-BF9D-4110-B512-7DD528DC7EDE} = {05AEF266-9A07-41B4-83C5-FC6632B27A0E}
 		{651ACC28-7A4C-41DB-8BAB-16830D4713BD} = {F23F1C19-7F80-41CC-86F7-F720C6D3C43B}
